refactor(FE): 코드 뷰어와 에디터 분리 (#146)
