+++ client/src/components/Code/Code.scss
@@ -1,10 +1,10 @@
 @import "../../constants/library";
 @import "@/styles/_global-styles";
-@import "@/components/CodeEditor/CodeTheme.scss";
+@import "@/components/Code/CodeTheme.scss";
 
 $line-width: 2.5rem;
 
-.code-editor {
+.code {
   position: relative;
   display: flex;
   box-sizing: border-box;
@@ -26,7 +26,7 @@ $line-width: 2.5rem;
     height: 100%;
     margin: 0;
     padding: 0.5rem;
-    font-size: 1rem;
+    font-size: 0.75rem;
     white-space: pre-wrap;
     &::-webkit-scrollbar {
       display: none;
