+++ client/src/components/Modal/WriteModal/WritingForm/WritingForm.scss
@@ -1,5 +1,5 @@
 @import "../SubmitModal/TitleInput/TitleInput";
-@import "src/components/CodeEditor/CodeEditor";
+@import "../../../Code/Code";
 @import "PrettifyButton/PrettifyButton";
 @import "SubmitButton/SubmitButton";
 @import "LanguageSelector/LanguageSelector";
